@@ -12,14 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.7",
-    "@angular/common": "~13.3.7",
-    "@angular/compiler": "~13.3.7",
-    "@angular/core": "~13.3.7",
-    "@angular/forms": "~13.3.7",
-    "@angular/platform-browser": "~13.3.7",
-    "@angular/platform-browser-dynamic": "~13.3.7",
-    "@angular/router": "~13.3.7",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@auth0/auth0-spa-js": "^2.0.4",
     "core-js": "^2.5.7",
     "json-server": "^1.0.0-alpha.23",
     "rxjs": "~6.5.3",
@@ -27,10 +28,10 @@
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.10",
-    "@angular/cli": "~13.3.7",
-    "@angular/compiler-cli": "~13.3.7",
-    "@angular/language-service": "~13.3.7",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/jasmine": "~3.10.1",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.2",
@@ -41,7 +42,7 @@
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "protractor": "^3.3.0",
+    "protractor": "^7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.3",
     "typescript": "~4.6.4"
